+++ src/command/contract/erc777_preset_fixed_supply.rs
@@ -1,10 +1,9 @@
-use super::{CallReturn, SendReturn};
-use crate::{
-  command::HexData,
-  openzeppelin::contracts::token::erc777::presets::erc777_preset_fixed_supply::Contract,
-};
+use super::{dump, CallReturn, SendReturn};
+use crate::openzeppelin::contracts::token::erc777::presets::erc777_preset_fixed_supply::Contract;
+use crate::{command::HexData, WrappedResult};
 use ethcontract::{
   dyns::{DynDeployBuilder, DynWeb3},
+  futures::StreamExt,
   Address, U256,
 };
 use structopt::StructOpt;
@@ -237,5 +236,118 @@ impl SendCommand {
       Self::TransferFrom { holder, recipient, amount }
         => contract.transfer_from(holder, recipient, amount).into(),
     }
+  }
+}
+
+#[derive(Debug, StructOpt)]
+#[structopt(rename_all = "verbatim")]
+pub enum EventsCommand {
+  #[structopt(about = "Any events for this contract")]
+  All,
+  #[structopt(
+    about = "Emitted when the allowance of an account is set for another.",
+    long_about = "Emitted when the allowance of a <spender> for an <owner> is set by a call to {approve}. <value> is the new allowance."
+  )]
+  Approval,
+  #[structopt(
+    about = "Emitted when an account is authorized as an operator.",
+    long_about = ""
+  )]
+  AuthorizedOperator,
+  #[structopt(about = "Emitted when a token is destroyed.")]
+  Burned,
+  #[structopt(about = "Emitted when a token is created.")]
+  Minted,
+  #[structopt(about = "Emitted when an account is revoked as an operator.")]
+  RevokedOperator,
+  #[structopt(about = "Emitted when tokens are moved from one account to another using {Send}.")]
+  Sent,
+  #[structopt(
+    about = "Emitted when tokens are moved from one account to another using {Transfer}.",
+    long_about = "Emitted when <value> tokens are moved from one account (<from>) to another (<to>)."
+  )]
+  Transfer,
+}
+
+impl EventsCommand {
+  pub async fn execute(self, web3: &DynWeb3, address: Address, stream: bool) -> WrappedResult<()> {
+    let contract = Contract::at(web3, address);
+
+    if stream {
+      match self {
+        Self::All => contract.all_events().stream().for_each(dump::stream).await,
+        Self::Approval => {
+          contract
+            .events()
+            .approval()
+            .stream()
+            .for_each(dump::stream)
+            .await
+        }
+        Self::AuthorizedOperator => {
+          contract
+            .events()
+            .authorized_operator()
+            .stream()
+            .for_each(dump::stream)
+            .await
+        }
+        Self::Burned => {
+          contract
+            .events()
+            .burned()
+            .stream()
+            .for_each(dump::stream)
+            .await
+        }
+        Self::Minted => {
+          contract
+            .events()
+            .minted()
+            .stream()
+            .for_each(dump::stream)
+            .await
+        }
+        Self::RevokedOperator => {
+          contract
+            .events()
+            .revoked_operator()
+            .stream()
+            .for_each(dump::stream)
+            .await
+        }
+        Self::Sent => {
+          contract
+            .events()
+            .sent()
+            .stream()
+            .for_each(dump::stream)
+            .await
+        }
+        Self::Transfer => {
+          contract
+            .events()
+            .transfer()
+            .stream()
+            .for_each(dump::stream)
+            .await
+        }
+      }
+    } else {
+      match self {
+        Self::All => dump::query(contract.all_events().query().await?),
+        Self::Approval => dump::query(contract.events().approval().query().await?),
+        Self::AuthorizedOperator => {
+          dump::query(contract.events().authorized_operator().query().await?)
+        }
+        Self::Burned => dump::query(contract.events().burned().query().await?),
+        Self::Minted => dump::query(contract.events().minted().query().await?),
+        Self::RevokedOperator => dump::query(contract.events().revoked_operator().query().await?),
+        Self::Sent => dump::query(contract.events().sent().query().await?),
+        Self::Transfer => dump::query(contract.events().transfer().query().await?),
+      }
+    }
+
+    Ok(())
   }
 }